refactor: split image search component
